Remove EvalUtil and move evaluation method into System
